I have been working hard. Changed the name of all "pick" related files and directories to "draw" related stuff for the final project. I also have been trying to use c++ nodes to access information produced by the franka library not accesible using python nodes. I think I am on the bring of being successful at this, but I have been working on it for many hours so I want to commit.
I also implemented code that can (hopfully)
open and close the gripper to specific distances,
whereas in our hw3 code all we could do was open
it all the way, or close it all the way. I think
that we were doing everything wrong. We were telling
the gripper to move to a distance of 0.08, that is,
move the pincers a distance of 0.08m apart. Funny
thing is, I think the grippers can only move 0.035m
apart. So whatever we were getting it to do is
undocumented behavior.