Added an alphabet_interfaces package.
Implemented a dictionary to handle which letters
we want to draw. You can do something like
load_moves('a') in the draw.py file to load
in all the points as poses to the queue.

Also I removed all the force control
testing stuff I had created an was working on from this branch. Not needed
and just taking up space, potentially confusing teammates.