Force control works.
Force control is working! What we are doing is
measuring the joint torque in joint 6, which is
holding up panda_link7 and panda_hand and the
gripper attachment. This doesn't change as long as we don't move
joint_6, so we're able to use this force as an offset for when
the robot is not running into the whiteboard. Once the robot
starts to run into the whiteboard, we'll be able to figure out the
force in the x-direction of the {s} frame using the torque in
joint_6, minus the offset we took before.