Force control works... maybe.
I figured out that in order to discretize the JointTrajectories
I take from planning a path, I have to go in and modify the
time_from_start parameter of the JointTrajectoryPoint to be
100000000 nanoseconds, instead of the original value given to it
during the initial path planning. I also made the period of my
timer_callback in the send_trajectories.py node to be 0.1, which
matches the time_from_start. It seems like the trajectories
are then fed to the robot at around the rate it was receiveing
them when we were using ExecuteTrajectory Goals. This worked well
in simulation.