Force control is working!!
It's official folks. I spent like four hours fighting
with compute_cartesian_path today, only to learn that
the robot I was working on just didn't work with it. Once
I switched robots it works. I still need to integrate Anunya's
code into mine to do the tf transforms, but essentially we have
gotten to a point where we can write and write with force control.