Significant restructuring well underway.
I restructed the draw.py file heavily. Essentially,
the file will take in two types of messages from the brain.py
file:

1. moveit motion planning messages for large moves, like
   doing a 180 degree turn.
2. cartesian motion planning messages for approaching the board
   and drawing letters.

The moveit planning messages are very simple to deal with. The cartesian
ones are harder. Here are the details

* The cartesian planning messages should have two components:
    1. a list of y and z coodinates that are for drawing the letters
    2. a start position of the letter, but only using y and z coordinates.

* The x coordinate of the points we plan for is special. It has to be
  determined by the routine of pressing the marker into the board.
  Therefore, it makes sense to figure out the x coordinate of the letter
  waypoint inside the draw.py file, and inside the force control logic.

* Everytime the draw.py file receives a message on the /cartesian_mp
  topic, it will do the following things
    1. move to the y and z coordinates of the start position, while
       maintaining the same x coordinate. This is because not doing this
       will interfere with out ability to touch the marker to the board
       at precise locations. I can't think of a way to put this super
       concisely right now, so if you have questions please come ask me.
       Just know that this is necessary.
    2. Move to the board. This will be a waypoint that is behind the
       board to ensure the robot collides with the surface of the
       board and stops its motion due to the force control. After the
       force control has stopped the motion of the end effector, the x
       coordinate will be recorded, and it will then be used to generate
       a list of poses to be planned for with the /compute_cartesian_path
       service. This will be what actually draws the letter.

I also created a mock brain.py file, although it's not close to complete
yet. It's not even in my setup.py file yet. This will eventually have
its own state machine that controls when messages get sent to the draw.py
function. It also defines the home position of the robot.

The home position of the robot will be a point close to the board but
not touching the board. After drawing a letter, the robot will return to
this position every time. This helps with the flow of the robots movement.

I have draw multiple possible flow charts for the robot's logic, and can
send pictures of them and explain them if any of you are interested.