Changed MoveJointState again.
The MoveJointState.srv file now has two different arguements.
One is a list of joint names that you would like to plan new
position for, and the second is a list of joint positions associated
with each of the joint names you previously entered. You can plan
positions for multiple joints at once this way, and the joints you
do not plan for will simply retain their current positions.

This goes without saying, but all positions are in radians.