Measuring accurate offset torque from gripper mass
Using the mass of the panda gripper and the location of the center of
mass of the panda gripper, I am successfully calculating the torque
in panda_joint6 due to the mass of the gripper. This happens correctly
at all orientations of panda_link6, even when it's rotated around its
x-axis. When panda_link6 is rotated around its x axis, panda_joint6
doesn't have to support the full weight of the gripper because some of
the force acts axially to the joint itself.