When force threshold is reached, plan is replanned
if the force threshold is reached int he send_trajectories node, then
it will replan all of the remaining trajectories in its queue to be
trajectories starting from the current point where the end-effector is.

I created a new custom service to do this, "replan", which should be as
fast as possible. It bypasses the state machines of both draw.py and
send_trajectories.py.