New force control logic
After spending all night trying to get the force control to work, I
sat down and tried to rethink everything I'm doing. This force control
code should, theoretically, be pretty robust and ready for testing on
the robot.

If this does not work, perhaps a next step could be to change the
frequency of the nodes to be 1000hz. This would allow me to take a
moving average of the force at the end effector, which would reduce
jitter.